fix: set id to unsigned bigint and fix calculate_status

+++ forlorn/src/infrastructure/redis/publish/announce.rs
@@ -1,7 +1,7 @@
 use super::publish;
 use crate::infrastructure::redis::RedisConnectionManager;
 
-pub async fn announce(redis: &RedisConnectionManager, score: i32) -> anyhow::Result<()> {
+pub async fn announce(redis: &RedisConnectionManager, score: u64) -> anyhow::Result<()> {
     publish(redis, "refx:announce", &score.to_string()).await?;
 
     Ok(())
